requirements.txt
# For python 3.11.9
fastapi==0.104.1
uvicorn==0.24.0
spotipy==2.23.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart==0.0.6
httpx==0.25.2
# scikit-learn>=1.3.0  # Temporarily disabled due to scipy issues
# pandas==2.1.3
# scikit-learn==1.3.2
# numpy==1.25.2